refactor auth route to controller and router
