+++ backend/package.json
@@ -8,6 +8,9 @@
     "@types/cookie-parser": "^1.4.2",
     "@types/jsonwebtoken": "^8.5.0",
     "@types/node": "^8.0.29",
+    "@types/swagger-jsdoc": "^6.0.0",
+    "@types/swagger-ui-express": "^4.1.2",
+    "concurrently": "^6.0.0",
     "nodemon": "^2.0.7",
     "ts-node": "3.3.0",
     "typescript": "3.3.3333"
@@ -30,6 +33,9 @@
     "multer": "^1.4.2",
     "pg": "^8.4.0",
     "reflect-metadata": "^0.1.10",
+    "swagger-jsdoc": "^6.1.0",
+    "swagger-ui-express": "^4.1.6",
+    "tsoa": "^3.6.1",
     "typeorm": "0.2.31",
     "typeorm-seeding": "^1.6.1"
   },
