--- backend/src/ormconfig.json
@@ -1,19 +1,0 @@
-{
-  "type": "postgres",
-  "host": "localhost",
-  "port": 5432,
-  "username": "test",
-  "password": "test",
-  "database": "readit",
-  "synchronize": false,
-  "logging": true,
-  "entities": ["entities/**/*{.ts,.js}"],
-  "migrations": ["migrations/*{.ts,.js}"],
-  "subscribers": ["subscribers/*{.ts,.js}"],
-  "seeds": ["seeds/**/*{.ts,.js}"],
-  "cli": {
-    "entitiesDir": "entity",
-    "migrationsDir": "migrations",
-    "subscribersDir": "subscribers"
-  }
-}
